implement serialize for api data sending
